update graph libraries

+++ verify/geometry/convex_is_convex.test.cpp
@@ -1,13 +1,13 @@
 #define PROBLEM "http://judge.u-aizu.ac.jp/onlinejudge/description.jsp?id=CGL_3_B"
 
-#include <bits/stdc++.h>
+#include <iostream>
 
-#include "geometry/polygon_is_convex.hpp"
+#include "geometry/polygon.hpp"
 
 int main() {
     int N;
     std::cin >> N;
-    Polygon<double> P(N);
+    Polygon<long long> P(N);
     std::cin >> P;
     bool ans = polygon_is_convex(P);
     std::reverse(P.begin(), P.end());
